Aula 3 completa

+++ src/http/routes/vote-on-poll.ts
@@ -2,6 +2,8 @@ import { randomUUID } from 'node:crypto';
 import { FastifyInstance } from 'fastify';
 import { prisma } from '../../lib/prisma';
 import z from 'zod';
+import { redis } from '../../lib/redis';
+import { voting } from '../../utils/voting-pub-sub';
 
 export async function voteOnPoll(app: FastifyInstance) {
     app.post('/polls/:pollId/votes', async (request, reply) => {
@@ -14,7 +16,7 @@ export async function voteOnPoll(app: FastifyInstance) {
         });
 
         const { pollId } = voteOnPollParams.parse(request.params);
-        const { pollOptionId } = voteOnPollBody.parse(request.body); // Valida e retorna request.body
+        const { pollOptionId } = voteOnPollBody.parse(request.body);
 
         let { sessionId } = request.cookies;
 
@@ -28,18 +30,23 @@ export async function voteOnPoll(app: FastifyInstance) {
                 }
             })
 
-            if (userPreviewsVoteOnPoll) {
-                if (userPreviewsVoteOnPoll.pollOptionId !== pollOptionId) {
-                    await prisma.vote.delete({
-                        where: {
-                            id: userPreviewsVoteOnPoll.id
-                        }
-                    })
-                } else {
-                    return reply.status(400).send({
-                        message: 'You already voted on this poll.'
-                    })
-                }
+            if (userPreviewsVoteOnPoll && userPreviewsVoteOnPoll.pollOptionId !== pollOptionId) {
+                await prisma.vote.delete({
+                    where: {
+                        id: userPreviewsVoteOnPoll.id
+                    }
+                });
+
+                const votes = await redis.zincrby(pollId, -1, userPreviewsVoteOnPoll.pollOptionId);
+
+                voting.publish(pollId, {
+                    pollOptionId: userPreviewsVoteOnPoll.pollOptionId,
+                    votes: Number(votes)
+                });
+            } else if (userPreviewsVoteOnPoll) {
+                return reply.status(400).send({
+                    message: 'You already voted on this poll.'
+                })
             }
         }
 
@@ -48,7 +55,7 @@ export async function voteOnPoll(app: FastifyInstance) {
 
             reply.setCookie('sessionId', sessionId, {
                 path: '/',
-                maxAge: 60 * 60 * 24 * 30, // 30 dias
+                maxAge: 60 * 60 * 24 * 30,
                 signed: true,
                 httpOnly: true
             })
@@ -60,7 +67,14 @@ export async function voteOnPoll(app: FastifyInstance) {
                 pollId,
                 pollOptionId
             }
-        })
+        });
+
+        const votes = await redis.zincrby(pollId, 1, pollOptionId);
+
+        voting.publish(pollId, {
+            pollOptionId,
+            votes: Number(votes)
+        });
 
         return reply.status(201).send();
     });
